json: fix some warnings and errors building tests.
CMakeLists.txt: fix missing adfun2json.cpp
adfun2json.hpp: fix type conversion warning.
json2adfun.hpp: fix type conversion warning.
19.omh: user's view of recent changes.
example_list.omh: change contents to have both tag and title.
example.omh: put list of examples higher in list of sections

+++ include/cppad/core/json/adfun2json.hpp
@@ -168,7 +168,7 @@ std::string CppAD::ADFun<Base,RecBase>::json(void)
         par2node[i_par] = ++dynamic_node;
         //
         // number of arguments for operators with fixed number of arguments
-        size_t n_arg = num_arg_dyn(op);
+        size_t n_arg = size_t( num_arg_dyn(op) );
         CPPAD_ASSERT_UNKNOWN( n_arg <= 2 );
         //
         // arguments in graph node space
